trigger prereqs class with title
Previously, Div creation was triggered by a Blockquote that began with a
header that contained a single class that was in a list of special
classes.

Now, Div creation is *also* triggered by a Blockquote that begins with a
header that has it's text contained in a list of special titles. The
title is used to lookup an appropriate class to give the Div.

In particular, 'prerequesites' is a special title, giving the class
'prereq'.

This input:

    > ## Prerequesites
    >
    > A short paragraph describing what learners need to know before
    > tackling this lesson.

will trigger this output:

    <div id="prerequisites" class="prereq">
    <h2>Prerequisites</h2>
    <p>A short paragraph describing what learners need to know before
    tackling this lesson.</p>
    </div>