swc.css: Add striped rows to tables
Now the style is matching closely `.table-bordered` from Bootstrap.
I decided to leave the header row background blank because it was too
close in color to `#F9F9F9` - our new rule for even rows in the table.

I also increased horizontal padding by 5px on both sides, because now
that we have vertical borders it was too cramped.

Additionaly I introduced borders for every cell in the table.