Modify the last two paragraphs
The last paragraph was confussing and had some inaccuracies. This version brings two new paragraphs, instead. The first defines what a commit as a intended set of changes to the files, and a repo as the history of commits and their metadata. It links to the glossary entries and doesn't mention Git at all.

The second paragraph is a callout that mentions other VCS and stresses the advantage of DVCS, like Mercurial and Git.

I removed the reference to the `.git` folder because this is mentioned in a callout in the "Tracking changes topic".