11-licensing: Tighten opening paragraph
No changes to the thrust, but this commit cleans up a few details.

Once we've introduced "creative work" and provided some examples,
stick to the more general term for the rest of the paragraph (instead
of shifting between "creative work", "code", "source code", etc.).

Drop the "has not waived" sentence.  We just said it's copyright, and
the distinction between licencing and "advertis[ing] as freely
available" isn't clear to me.

In the "reusing" sentence, the threat is the copyright holders, which
may not be the original authors (the authors may have assigned their
copyright to another party).  And the unlicensed infringement will
have always been illegal, it doesn't become illegal when the copyright
holders start enforcing their copyright.